pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "iskra-engine"
version = "0.1.0"
description = "RAG 引擎，提供对内的 HTTP API：基于 PostgreSQL/pgvector 的检索与 LLM 对话调用；不负责会话持久化等外层 Web 服务。"
readme = "README.md"
# 新版 Python 没有预编译轮子
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.128.0",
    # umbrella ≥ 0.14.21：默认依赖比 0.14.13 更精简（CLI/readers/cloud 按需单独安装 integration 包）
    "llama-index>=0.14.21",
    # umbrella = 0.14.13 时，banks（llama-index-core 依赖）在 griffe 1.x 下会破坏 import，需要锁到 0.x
    # "griffe>=0.49,<1",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    # llama-cpp-python 0.3.19 — 仅 CPU 预编译；放弃 GPU 另装 wheel
    # Windows ：abetlen Release 直链（cp312 · win_amd64）
    "llama-cpp-python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.19/llama_cpp_python-0.3.19-cp312-cp312-win_amd64.whl ; sys_platform == 'win32'",
    # Linux   ：同版本，依赖下方 [tool.uv] extra-index（whl/cpu）解析预编译 wheel
    "llama-cpp-python==0.3.19 ; sys_platform == 'linux'",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[tool.setuptools.packages.find]
where = ["src"]

# 供 Linux 条 `llama-cpp-python==0.3.19` 从 abetlen 取 wheel，避免 PyPI 仅 sdist
[tool.uv]
extra-index-url = ["https://abetlen.github.io/llama-cpp-python/whl/cpu"]
